Fix bugs and change project structures
