+++ .gitignore
@@ -1,2 +1,2 @@
 .idea
-/*/__pycache__
+/*/*/__pycache__
