--- caller.py
@@ -1,19 +1,0 @@
-#!/usr/bin/env python
-
-import src.cli
-import src.deepin
-
-src.cli.color()
-src.cli.zsh()
-src.cli.vim()
-src.cli.pkgmgr()
-src.cli.sshd()
-src.cli.nfsrv()
-src.cli.misc()
-
-src.deepin.dde()
-#src.deepin.vbox
-src.deepin.deconf()
-src.deepin.ddeconf()
-src.deepin.prvconf()
-
