+++ src/install.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+
+from arch import cli, deepin
+
+cli.color()
+cli.zsh()
+cli.vim()
+cli.pkgmgr()
+cli.sshd()
+cli.nfsrv()
+cli.misc()
+
+deepin.dde()
+#deepin.vbox()
+deepin.deconf()
+deepin.ddeconf()
+deepin.prvconf()
+
